Add old project files
This is an old project that I created before I used good practices in my code.
It was able to take simple Coxeter-Dynkin diagrams and render them as 3d polytopes.
I think the concept is really cool, so I'd like to refactor it so that I don't feel ashamed about the code quality.